Move clip to common utilities
