+++ imod/common/utilities/clip.py
