+++ imod/common/utilities/regrid.py
@@ -21,7 +21,7 @@ from imod.mf6.auxiliary_variables import (
     expand_transient_auxiliary_variables,
     remove_expanded_auxiliary_variables_from_dataset,
 )
-from imod.mf6.utilities.clip import clip_by_grid
+from imod.common.utilities.clip import clip_by_grid
 from imod.schemata import ValidationError
 from imod.typing.grid import (
     GridDataArray,
