+++ imod/mf6/multimodel/modelsplitter.py
@@ -10,7 +10,7 @@ from imod.mf6.auxiliary_variables import (
 )
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.hfb import HorizontalFlowBarrierBase
-from imod.mf6.utilities.clip import clip_by_grid
+from imod.common.utilities.clip import clip_by_grid
 from imod.mf6.utilities.grid import get_active_domain_slice
 from imod.mf6.wel import Well
 from imod.typing import GridDataArray
