+++ imod/tests/test_mf6/test_utilities/test_clip_utilities.py
@@ -11,7 +11,7 @@ from shapely.testing import assert_geometries_equal
 import imod
 from imod.common.interfaces.ipackagebase import IPackageBase
 from imod.mf6 import HorizontalFlowBarrierResistance
-from imod.mf6.utilities.clip import clip_by_grid
+from imod.common.utilities.clip import clip_by_grid
 from imod.mf6.utilities.grid import broadcast_to_full_domain
